fixes: 5465

+++ test/dialect/oracle/test_reflection.py
@@ -1,6 +1,7 @@
 # coding: utf-8
 
 
+from sqlalchemy import DOUBLE
 from sqlalchemy import exc
 from sqlalchemy import FLOAT
 from sqlalchemy import ForeignKey
@@ -803,7 +804,7 @@ class TypeReflectionTest(fixtures.TestBase):
         connection,
     ):
         specs = [
-            (DOUBLE_PRECISION(), FLOAT()),
+            (DOUBLE_PRECISION(), DOUBLE()),
             # when binary_precision is supported
             # (DOUBLE_PRECISION(), oracle.FLOAT(binary_precision=126)),
             (BINARY_DOUBLE(), BINARY_DOUBLE()),
